fix: Pass nonce and endpoint down to verification
Addressing #1163, we are now saving the nonce for presentations and
passing it along with the config endpoint down to the presentation
verification logic. This should begin verifying the KB-JWT, but I'm a
bit worried that the code may be fragile. I do not know the spec well
enough to know if the KB-JWT is required when the verifier passes a
nonce, or if it is completely optional.

This change effectively makes it required 100% of the time for
presentations.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>