feat: Pass full presentation record to verifiers
Instead of passing down the nonce & aud parameters to the
verify_presentation method, pass down the entire presentation record
instead.

Since we're retrieving the AUD field from the profile's config, let's
just pull it out when we need it.

Also realized that verify_credential isn't even called from the route,
and so adding the presentation/nonce info may have broken whatever code
calls it (specifically in pex.py, which I did not modify and don't
understand how it works). Therefore, it was safer to remove it from the
signature.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>